requirements.txt
contourpy==1.3.2
cycler==0.12.1
faiss-cpu==1.11.0
fonttools==4.58.4
hnswlib==0.8.0
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
